refactor: 발신 데이터, 리스너 타입 변경
