+++ perfume-api/src/main/java/io/perfume/api/notification/adapter/port/in/dto/DeleteNotificationResponseDto.java
@@ -1,3 +1,3 @@
-package io.perfume.api.notification.adapter.port.in;
+package io.perfume.api.notification.adapter.port.in.dto;
 
 public record DeleteNotificationResponseDto(Long id) {}
