+++ perfume-api/src/main/java/io/perfume/api/review/application/facade/dto/ReviewCommentEvent.java
@@ -1,0 +1,3 @@
+package io.perfume.api.review.application.facade.dto;
+
+public record ReviewCommentEvent(Long reviewId, Long receiveUserId) {}
